backend/requirements.txt
fastapi==0.115.0 
uvicorn==0.30.6 
sqlalchemy==2.0.35 
# psycopg2-binary==2.9.9 
pydantic==2.9.2 
python-multipart==0.0.9 
requests==2.32.3 
python-dotenv==1.0.1
